Updated -set and -reset to handle scenes settings
An extra argument after `-set` or `-reset` allows changing scene specific
settings. For example, setting the skymap for a single scene can be done like
this:

    chunky -set sky.skymap /path/to/skymap.jpg scene-name

The skymap option can be reset by the command

    chunky -reset sky.skymap scene-name

The name of the setting can include periods, and in that case denotes a path
in the json object. The `sky.` part of the name indicates that the `skymap`
setting is a member of the `sky` object.

fixes #179 (github)