Fixed potential crash in launcher on clean install
If the launcher installed Chunky when there were no previous versions
available then it would crash after the install.

This meant that the first time you try to start Chunky with the new launcher
it would just close and it seems as if nothing happens. The second time it
would start Chunky correctly.