Various UI bug fixes
Editing the scene name should now take effect directly, even when the
scene is not currently rendering.

Pressing ENTER in the scene name field will save the scene.

Changing the number of render threads should now save the value.

Added a setting named "drawUnknownBlocks" which controls the debug block
rendering.

Loading a new scene with a different canvas size than the current one
should work better now. Previously it could cause a NullPointerException,
or draw the new canvas as a picture-in-picture in the old preview image.