Fix crash in the launcher for first-time setups
The launcher would crash for first-time setups because it relied on the
settings directory being initialized before the first-time setup window
had been shown.

Also removed dependence on PersistentSettings.settingsDirectory() in
the launcher because the value depends on the first-time setup, and
if it is called too soon it will always return null.

fixes #386 (github)