Extract JSON library to external project
The JSON library that was built as part of Chunky is now a separate library,
and included via a Maven dependency.
The JSON library lives here: https://github.com/llbit/jo-json

Moved the process for building of the release Jar to the Gradle build, in the
releaseJar task.  Only version information is processed by the old
ReleaseBuilder tool.