Refactor and document the octree implementations (#615)
* refactored PackedOctree intersection

* Fixed enterOctree()

The code that was put there required several iterations to reach the
octree. I replaced it with a single AABB intersection.

* Copied the changes over to NodeBasedOctree

* Fixed a bug with enterOctree()

The normal direction when hitting a z side was computed with a value on
the y axis.

* Added a short description to enterOctree()

* Applied same refactoring to exitWater()

* Added more documentation to octree implementations

Added more comments explaining what's happening in octree intersection
functions (enterBlock() and exitWater()).
Wrote proper JavaDoc for enterOctree().

* Added more documentation to octree intersection

I added JavaDoc for enterBlock() and exitWater(). I also added a short
description of how the octrees are intersected inside those functions.

* Fix missing word in JavaDoc for enterOctree()