Rewrote how render tasks are fed to workers
The RenderManager was doing a lot of manual synchronization tailored to
feeding the same jobs for each frame. This rewrite uses Java
synchronization primitives and allows more freedom in what and how many
jobs are given per frame. This is precursor work to adaptive sampling
that will generate different jobs each frame.

There may be a performance hit, will need to monitor and investigate.