Some of the suggested modifications by leMaik
Minor changes like rename and slight logic changes, and:

Now using the `int[][] spp` to store spp for each pixel; this will increase memory use a bit, but also opens the door for not rendering the whole canvas at the same time. This is not fully implemented yet, as Scene#spp is used pretty much everywhere still.