Increase Ray Precision
No noticable speed cost of increased precision, while reducing rays from hitting the "seams" between blocks, or leaking through diagonals. (Not a true fix, but reduces the problem by 99%.) These seams are really only noticible if you zoom in to the edge of a block now, and leaked rays are are fewer and would make much less of an impact (1/100 as many as the few that go through now should make them nigh invisible).

(And while I'm at this super short PR, may as well suppress the only warning, which due to how it is currently used, will never cause crash.)