Fix various octree bugs (#933)
* Fix an of by one error where, if yMax was a multiple of 16, an additional iteration was done, inserting an additional empty cube in the octree.
If the octree was small enough that this additional cube was out of bounds, it wrapped around and overwrote a valid cube, leading to the bottoms layers being replaced with air

* Change the bound check against yClipMax to be exclusive when loading entities to be more consistent with the bound check against yMax when loading blocks.
Change the bound check from yClip(Min|Max) to y(Min|Max) when loading block entity so the test is the same as when loading blocks because those entities are related to blocks. For the other entities, still using yClip(Min|Max) is desirable to be able to load entities that are below 0/higher than 256

* Only try to merge node starting at the depth of the inserted cube (because deeper than that, inside the cube, the parents have not been set and it would be useless anyway due to how the cube insertion is done).
This was the cause of the StackOverflow in finalizeNode, due to a cycle in treeData, due to bad merges

* Use yMin as the y coordinate of the origin so the coordinate inside the octree always start at 0 even for negative of greater than 0 y values.
As for x and z, the y coordinate is forced to be a multiple of 16 to correctly work with the way the chunks are loaded in the octree.
I think there is no need for additional changes to make entities (and emitter position) align with the loading chunks.
I think that this change is technically not backward compatible because the way the origin is computed is different, but it only breaks compatibility with snapshot scene because scene created in stable release are already treated differently

* Change the nodeEquals method to not return true when comparing two branch nodes.
It lead to wrongful merge when a node had 8 children because every node compared equals even though they were different.